+++ scripts/delete_minikube.sh
@@ -1,0 +1,4 @@
+. /sv/scripts/errorHandler.sh
+. /sv/scripts/requireRoot.sh
+
+sudo -H -u vagrant minikube delete --all --purge