+++ scripts/install_kubectl.sh
@@ -1,10 +1,9 @@
 . /sv/scripts/errorHandler.sh
+. /sv/scripts/variables.sh
 
-kubectl_version="v1.21.14" # previous version v1.17.4
-current_kubectl_version=$(kubectl version --client --short 2> /dev/null || true)
-kubectl_version_expected=$"Client Version: $kubectl_version"
+current_kubectl_version=$(kubectl version --client -o json | jq -r .clientVersion.gitVersion 2> /dev/null || true)
 
-if [ "$current_kubectl_version" != "$kubectl_version_expected" ]; then
+if [ "$current_kubectl_version" != "$kubectl_version" ]; then
 	apt-get update
 	apt-get install -y curl