+++ scripts/requireRoot.sh
@@ -1,0 +1,4 @@
+if [ `whoami` != "root" ]; then
+	echo "Script must be executed as root"
+	exit 1
+fi