+++ changelog.md
@@ -1,5 +1,11 @@
 # sv-kubernetes changelog
 
+## 6/8/2023
+* Minikube updated to `1.30.1`.
+* Kubectl updated to `1.26.5`.
+* Implements packer build system.
+* Minikube now runs using `driver=docker`.
+
 ## 9/29/2020
 * Minikube updated to 1.13.0 from 1.3.0
 * The vagrantfile will now allocate a 100gb disk to grant more space to applications than need to house databases.